Исправления замечаний наставника модуля 7
