+++ js/main.js
@@ -1,6 +1,10 @@
-import './data.js';
-import './render-photos.js';
-import './big-photo.js';
 import './img-upload.js';
-import './validation-form.js';
-import './api.js';
+import {drawPhotos} from './render-photos.js';
+import {initializeForm} from './form.js';
+import {getData} from './api.js';
+import {showAlert} from './messages.js';
+
+
+getData(drawPhotos, showAlert);
+initializeForm();
+
